Add files via upload
A 3-to-8 decoder takes a 3-bit binary input and generates 8 outputs, with only one 
output active (logic high) at a time. The output corresponding to the binary input is set to high (1), 
while all other outputs remain low (0). The decoder essentially performs a 1-of-8 selection based on 
the 3-bit input.