Add files via upload
In digital signal processing, an additional bit either 0 or 1 is added to the original binary or digital code to detect and correct any kind of errors in the data that can occur during transmission. This additional bit is called a parity bit.
The addition of a parity bit to the original digital code makes the total number of 1s in the code either even or odd. Thus, on the basis of number 1s in the data, the parity can be classified into two types namely, even parity and odd parity.

First, a parity generator reads the input data and calculates the parity bit accordingly. Once the parity bit is generated, it is added to the original data code. This gives an output code which is the original code with a newly generated parity bit.

The parity generator is used at the transmitter end and generate and add a parity bit the original code before transmission.